+++ js/constants.js
@@ -1,0 +1,30 @@
+const GALLERY_LENGTH = 25;
+const DESCRIPTIONS = [
+  'Мои собаки',
+  'Солнечная погода',
+  'На работе',
+  'Отдых',
+  'Xbox one X',
+  'В гостях у друзей',
+  'Учим Frontend',
+  'Москва',
+  'Новый год 2022',
+  'Учим Backend',
+  'Сноубординг',
+  'Сочи',
+  'Бетанкур',
+  'Мои мыши',
+  'Бар',
+  'Собираю мебель',
+  'Вечер за сериалом',
+  'Участвую в конурсе',
+  'Новая работа',
+  'Продаю PS4',
+  'Новое хобби',
+  'Кросс',
+  'С друзьями в зале',
+  'Велоезда',
+  'Мой первый обзор на фильм'
+];
+
+export {GALLERY_LENGTH, DESCRIPTIONS};
